+++ contracts/factory/src/tests/config.rs
@@ -1,10 +1,11 @@
 use super::setup::{
-    deploy_factory_contract, install_lp_contract, install_multihop_wasm, install_stable_lp,
-    install_stake_wasm, install_token_wasm, lp_contract,
+    deploy_factory_contract, install_latest_factory, install_lp_contract, install_multihop_wasm,
+    install_stable_lp, install_stake_wasm, install_token_wasm, lp_contract,
 };
 use crate::{
     contract::{Factory, FactoryClient},
     error::ContractError,
+    storage::ADMIN,
     tests::setup::{generate_lp_init_info, install_and_deploy_token_contract, stable_lp},
 };
 
@@ -521,6 +522,25 @@ fn update_wasm_hashes() {
 }
 
 #[test]
+fn migrate_admin_key() {
+    let env = Env::default();
+    env.mock_all_auths();
+    env.cost_estimate().budget().reset_unlimited();
+
+    let admin = Address::generate(&env);
+
+    let factory = deploy_factory_contract(&env, Some(admin.clone()));
+
+    factory.migrate_admin_key();
+
+    let after_migration: Address = env.as_contract(&factory.address, || {
+        env.storage().instance().get(&ADMIN).unwrap()
+    });
+
+    assert_eq!(admin, after_migration);
+}
+
+#[test]
 fn update_multihop_addr() {
     let env = Env::default();
     env.mock_all_auths();
@@ -550,6 +570,22 @@ fn update_multihop_addr() {
 }
 
 #[test]
+fn test_simple_update() {
+    let env = Env::default();
+    env.mock_all_auths();
+    env.cost_estimate().budget().reset_unlimited();
+
+    let admin = Address::generate(&env);
+
+    let factory = deploy_factory_contract(&env, Some(admin.clone()));
+
+    let new_wasm_hash = install_latest_factory(&env);
+    let latest_wasm = install_stable_lp(&env);
+
+    factory.update(&new_wasm_hash, &latest_wasm);
+}
+
+#[test]
 fn update_lp_token_decimals() {
     let env = Env::default();
     env.mock_all_auths();
@@ -576,6 +612,21 @@ fn update_lp_token_decimals() {
     let updated_token_decimals = factory.get_config().lp_token_decimals;
 
     assert!(updated_token_decimals == new_token_decimals);
+}
+
+#[test]
+fn test_query_version() {
+    let env = Env::default();
+    env.mock_all_auths();
+    env.cost_estimate().budget().reset_unlimited();
+
+    let admin = Address::generate(&env);
+
+    let factory = deploy_factory_contract(&env, Some(admin.clone()));
+
+    let expected_version = env!("CARGO_PKG_VERSION");
+    let version = factory.query_version();
+    assert_eq!(String::from_str(&env, expected_version), version);
 }
 
 #[test_case(0u32; "should fail with 0")]