+++ contracts/multihop/src/contract.rs
@@ -394,6 +394,7 @@ impl Multihop {
     }
 
     #[allow(dead_code)]
+    #[cfg(not(tarpaulin_include))]
     //TODO: Remove after we've added the key to storage
     pub fn add_new_key_to_storage(env: Env) -> Result<(), ContractError> {
         env.storage().persistent().set(&MULTIHOP_KEY, &true);