+++ contracts/pool_stable/src/contract.rs
@@ -997,13 +997,6 @@ impl StableLiquidityPool {
         env.events()
             .publish(("initialize", "XYK LP token_b"), token_b);
     }
-    #[allow(dead_code)]
-    pub fn update(env: Env, new_wasm_hash: BytesN<32>) {
-        let admin = get_admin_old(&env);
-        admin.require_auth();
-
-        env.deployer().update_current_contract_wasm(new_wasm_hash);
-    }
 
     #[allow(dead_code)]
     pub fn query_version(env: Env) -> String {
@@ -1012,6 +1005,7 @@ impl StableLiquidityPool {
 
     #[allow(dead_code)]
     //TODO: Remove after we've added the key to storage
+    #[cfg(not(tarpaulin_include))]
     pub fn add_new_key_to_storage(env: Env) -> Result<(), ContractError> {
         env.storage().persistent().set(&STABLE_POOL_KEY, &true);
         Ok(())