+++ contracts/stake/src/storage.rs
@@ -135,6 +135,7 @@ pub mod utils {
         );
     }
 
+    #[cfg(not(tarpaulin_include))]
     pub fn _save_admin(e: &Env, address: &Address) {
         e.storage().instance().set(&ADMIN, &address);
         e.storage()
@@ -153,6 +154,7 @@ pub mod utils {
         admin
     }
 
+    #[cfg(not(tarpaulin_include))]
     pub fn _get_admin(e: &Env) -> Address {
         e.storage()
             .instance()