+++ contracts/trader/src/contract.rs
@@ -451,6 +451,7 @@ impl Trader {
     }
 
     #[allow(dead_code)]
+    #[cfg(not(tarpaulin_include))]
     //TODO: Remove after we've added the key to storage
     pub fn add_new_key_to_storage(env: Env) -> Result<(), ContractError> {
         env.storage().persistent().set(&TRADER_KEY, &true);