+++ contracts/vesting/src/tests/instantiate.rs
@@ -7,6 +7,8 @@ use crate::{
 };
 use curve::{Curve, SaturatingLinear};
 
+use super::setup::{install_latest_vesting, old_vesting};
+
 #[test]
 fn instantiate_contract_successfully() {
     let env = Env::default();
@@ -75,6 +77,14 @@ fn instantiate_contract_successfully() {
             }
         ]
     );
+
+    let config = vesting_client.query_config();
+    assert!(!config.is_with_minter);
+
+    let new_vesting_token = deploy_token_contract(&env, &Address::generate(&env));
+
+    vesting_client.update_vesting_token(&new_vesting_token.address);
+    vesting_client.update_max_complexity(&10);
 }
 
 #[should_panic(
@@ -143,4 +153,70 @@ fn create_schedule_panics_when_admin_has_no_tokens_to_fund() {
     );
 
     vesting_client.create_vesting_schedules(&vesting_schedules);
+}
+
+#[test]
+fn test_update_vesting() {
+    let env = Env::default();
+    env.mock_all_auths();
+
+    let admin = Address::generate(&env);
+
+    let token_client = deploy_token_contract(&env, &admin);
+
+    let vesting_token = old_vesting::VestingTokenInfo {
+        name: String::from_str(&env, "Phoenix"),
+        symbol: String::from_str(&env, "PHO"),
+        decimals: 6,
+        address: token_client.address.clone(),
+    };
+
+    let vesting_addr = env.register(old_vesting::WASM, ());
+    let old_vesting = old_vesting::Client::new(&env, &vesting_addr);
+
+    old_vesting.initialize(&admin, &vesting_token, &6);
+
+    let new_wasm_hash = install_latest_vesting(&env);
+    old_vesting.update(&new_wasm_hash);
+
+    let latest_vesting = VestingClient::new(&env, &old_vesting.address);
+    assert_eq!(
+        latest_vesting.query_token_info().address,
+        vesting_token.address
+    );
+}
+
+#[test]
+fn test_update_vesting_with_minter() {
+    let env = Env::default();
+    env.mock_all_auths();
+
+    let admin = Address::generate(&env);
+
+    let token_client = deploy_token_contract(&env, &admin);
+
+    let vesting_token = old_vesting::VestingTokenInfo {
+        name: String::from_str(&env, "Phoenix"),
+        symbol: String::from_str(&env, "PHO"),
+        decimals: 6,
+        address: token_client.address.clone(),
+    };
+
+    let vesting_addr = env.register(old_vesting::WASM, ());
+    let old_vesting = old_vesting::Client::new(&env, &vesting_addr);
+
+    let minter_addr = Address::generate(&env);
+    let minter_capacity = 6;
+    let minter_info = old_vesting::MinterInfo {
+        address: minter_addr.clone(),
+        mint_capacity: minter_capacity,
+    };
+
+    old_vesting.initialize_with_minter(&admin, &vesting_token, &6, &minter_info);
+
+    let new_wasm_hash = install_latest_vesting(&env);
+    old_vesting.update(&new_wasm_hash);
+
+    let latest_vesting = VestingClient::new(&env, &old_vesting.address);
+    assert_eq!(latest_vesting.query_minter().address, minter_addr);
 }